+++ src/main/java/Lorem_Software/Library_Maintenance_System/business/controller/UsuarioController.java
@@ -4,7 +4,6 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 
-// import org.hibernate.mapping.List;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -37,14 +36,14 @@ public class UsuarioController {
     private ReservaDAO reservaDAO;
 
     Usuario usuario;
-    
+
     public UsuarioController() {
-    	super();
-    	this.usuario = new Usuario(); 
+        super();
+        this.usuario = new Usuario();
     }
 
     @GetMapping("/AltaUsuario")
-    public String altaUsuario(Model model){
+    public String altaUsuario(Model model) {
         List<Usuario> listadoUsuarios = usuarioDAO.findAll();
         Collections.sort(listadoUsuarios, new UsuarioComparator());
 
@@ -57,7 +56,7 @@ public class UsuarioController {
     }
 
     @PostMapping("/AltaUsuario")
-    public String usuarioSubmit(@ModelAttribute Usuario usuario, Model model){
+    public String usuarioSubmit(@ModelAttribute Usuario usuario, Model model) {
         this.usuario = usuario;
         model.addAttribute("usuario", usuario);
         usuarioDAO.save(this.usuario);
@@ -65,21 +64,20 @@ public class UsuarioController {
     }
 
     @GetMapping("/AltaUsuario/edit/{IdUsuario}")
-    public String formularioEditarUsuario(@PathVariable("IdUsuario") long IdUsuario, Model model){
-        Optional<Usuario> usuario = usuarioDAO.findById(IdUsuario);
-        model.addAttribute("usuario", usuario);
+    public String formularioEditarUsuario(@PathVariable("IdUsuario") long idUsuario, Model model) {
+        Optional<Usuario> usuarioEditar = usuarioDAO.findById(idUsuario);
+        model.addAttribute("usuario", usuarioEditar);
         return "usuario/AltaUsuario";
     }
 
     @PostMapping("/AltaUsuario/update/{IdUsuario}")
-    public String actualizarUsuario(@PathVariable("IdUsuario") long IdUsuario, Usuario usuario){
+    public String actualizarUsuario(@PathVariable("IdUsuario") long idUsuario, Usuario usuario) {
         usuarioDAO.save(usuario);
         return "redirect:/AltaUsuario";
     }
 
-
     @GetMapping("/AltaUsuario/delete/{IdUsuario}")
-    public String deleteUsuario(@PathVariable("IdUsuario") long id){
+    public String deleteUsuario(@PathVariable("IdUsuario") long id) {
 
         suprimirAsociacionesPrestamo(usuarioDAO.findById(id));
         suprimirAsociacionesReserva(usuarioDAO.findById(id));
@@ -88,59 +86,61 @@ public class UsuarioController {
     }
 
     @PreRemove
-    private void suprimirAsociacionesPrestamo(Optional<Usuario> us){
-        Usuario u = us.get();
-        for(Prestamo p : u.getPrestamos()){
-            prestamoDAO.deleteById(p.getIdPrestamo());
+    private void suprimirAsociacionesPrestamo(Optional<Usuario> us) {
+        if(!us.isEmpty()){
+            Usuario u = us.get();
+            for (Prestamo p : u.getPrestamos()) {
+                prestamoDAO.deleteById(p.getIdPrestamo());
+            }
         }
     }
 
     @PreRemove
-    private void suprimirAsociacionesReserva(Optional<Usuario> us){
-        Usuario u = us.get();
-        for(Reserva r : u.getReservas()){
-            reservaDAO.deleteById(r.getIdReserva());
+    private void suprimirAsociacionesReserva(Optional<Usuario> us) {
+        if(!us.isEmpty()){
+            Usuario u = us.get();
+            for (Reserva r : u.getReservas()) {
+                reservaDAO.deleteById(r.getIdReserva());
+            }
         }
     }
 
-
     class UsuarioComparator implements java.util.Comparator<Usuario> {
-		@Override
-		public int compare(Usuario a, Usuario b) {
-			return a.getNombre().compareTo(b.getNombre());
-		}
-	}
+        @Override
+        public int compare(Usuario a, Usuario b) {
+            return a.getNombre().compareTo(b.getNombre());
+        }
+    }
 
+    public UsuarioDAO getUsuarioDAO() {
+        return usuarioDAO;
+    }
 
-	public UsuarioDAO getUsuarioDAO() {
-		return usuarioDAO;
-	}
+    public void setUsuarioDAO(UsuarioDAO usuarioDAO) {
+        this.usuarioDAO = usuarioDAO;
+    }
 
-	public void setUsuarioDAO(UsuarioDAO usuarioDAO) {
-		this.usuarioDAO = usuarioDAO;
-	}
+    public PrestamoDAO getPrestamoDAO() {
+        return prestamoDAO;
+    }
 
-	public PrestamoDAO getPrestamoDAO() {
-		return prestamoDAO;
-	}
+    public void setPrestamoDAO(PrestamoDAO prestamoDAO) {
+        this.prestamoDAO = prestamoDAO;
+    }
 
-	public void setPrestamoDAO(PrestamoDAO prestamoDAO) {
-		this.prestamoDAO = prestamoDAO;
-	}
+    public ReservaDAO getReservaDAO() {
+        return reservaDAO;
+    }
 
-	public ReservaDAO getReservaDAO() {
-		return reservaDAO;
-	}
+    public void setReservaDAO(ReservaDAO reservaDAO) {
+        this.reservaDAO = reservaDAO;
+    }
 
-	public void setReservaDAO(ReservaDAO reservaDAO) {
-		this.reservaDAO = reservaDAO;
-	}
+    public Usuario getUsuario() {
+        return usuario;
+    }
 
-	public Usuario getUsuario() {
-		return usuario;
-	}
-
-	public void setUsuario(Usuario usuario) {
-		this.usuario = usuario;
-	}
+    public void setUsuario(Usuario usuario) {
+        this.usuario = usuario;
+    }
 }
